+++ src/main/java/Game/Gui.java
@@ -30,14 +30,14 @@ public class Gui
     }
     public void render(ProgramContainer pc, Renderer r, Stats stats, Level level, double passedTime)
     {
-        r.drawStaticImage(pc, background.getImg(), background.getPosX(), background.getPosY());
+        r.drawStaticImage(pc, background.getImg(), (int)background.getPosX(), (int)background.getPosY());
         r.drawStaticText(pc, "HP: " + stats.getHp(), 912, 16, 0xFF000000, 3);
         r.drawStaticText(pc, "F: " + stats.getFire(), 912, 48, 0xFF000000, 3);
         r.drawStaticText(pc, "A: " + stats.getAir(), 912, 80, 0xFF000000, 3);
         r.drawStaticText(pc, "W: " + stats.getWater(), 912, 112, 0xFF000000, 3);
         r.drawStaticText(pc, "E: " + stats.getEarth(), 912, 144, 0xFF000000, 3);
         r.drawStaticText(pc, "Wave: " + (level.getCurrentWave() + 1) + "/" + level.getWavesAmount(), 912, 176, 0xFF000000, 4);
-        r.drawStaticImage(pc, nextWave.getImg(), nextWave.getPosX(), nextWave.getPosY());
+        r.drawStaticImage(pc, nextWave.getImg(), (int)nextWave.getPosX(), (int)nextWave.getPosY());
         r.drawStaticText(pc, "Next wave in:", 914, 204, 0xFF000000, 5);
         if(level.getCurrentWave() >= 0)
         {
