+++ src/main/java/Map/BackgroundTile.java
@@ -6,7 +6,7 @@ import engine.Renderer;
 
 public class BackgroundTile extends Tile
 {
-    public BackgroundTile(String path, int posX, int posY, int width, int height, int id)
+    public BackgroundTile(String path, double posX, double posY, int width, int height, int id)
     {
         super(path, posX, posY, width, height, id);
     }
